% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_meps.R
\name{dl_meps}
\alias{dl_meps}
\title{Download public use files from MEPS website}
\usage{
dl_meps(fname, dir = tempdir())
}
\arguments{
\item{fname}{name of public use file (standard format)}

\item{dir}{name of directory to store .ssp file. Defaults to tempdir() if none specified}
}
\value{
full path of data .ssp for import via read.xport
}
\description{
Helper function to download public use file from MEPS website. Can save .ssp file to a permanent directory.
}
