% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ssp.R
\name{download_ssp}
\alias{download_ssp}
\title{Download MEPS public use file}
\usage{
download_ssp(file, dir = "meps_data", force = F)
}
\arguments{
\item{file}{name of PUF file. Must be in standard format (e.g. 'h160g'). Can use the get_puf_names() function to convert year and file type to standard format.}

\item{dir}{local directory to store .ssp files. Defaults to 'meps_data' folder in current working directory. Will create folder if none exists.}

\item{force}{if TRUE, overwrites existing files. Otherwise, the function checks if file already exists in the specified directory.}
}
\description{
This function downloads SAS transport files (.ssp) from the MEPS website based on specified file and saves them in the specified folder. Requires internet connection.
}
\examples{
# Download .ssp file for 2015 full-year-consolidated data

download_ssp('h181')

# Download .ssp file for 2013 office-based event file, and store in 'mydata' folder

fname = get_puf_names(year=2013,type='OB')
download_ssp(fname,dir='mydata')
}
