% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ascii_info.R
\name{get_ascii_info}
\alias{get_ascii_info}
\title{Get ASCII info from Stata programming statements, for reading .dat files into R}
\usage{
get_ascii_info(filename)
}
\arguments{
\item{filename}{name of public use file. Must be in standard format (e.g. 'h160g'). Can use the get_puf_names() function to look up file name by year and type.}
}
\value{
list(start, end, names, types). List of ASCII file information, included variable start and end positions, names, and types
}
\description{
This function pulls information needed to load the MEPS PUFS in ASCII .dat file format into R. This program was built as a work-around for MEPS 2018 PUFs. Due to design changes in the survey instrument, the SAS transport files (.ssp) are output using the CPORT engine starting in 2018. Since this file type is not compatible with R, the ASCII (.dat) file must be used instead. This function pulls the variable names, types, and positions from the Stata programming statements. It is intended as a temporary work-around, until the R programming statements are released. Internet connection is required.
}
\examples{
## Import 2018 dental file (H206b)

# Pull ASCII file info from MEPS website.
dn_info <- get_ascii_info('h206b')

# Import data using read_fwf function (must save .dat file to local directory first)
meps_dat <- read_fwf("C:/MEPS/h206b.dat",
  col_positions = 
    fwf_positions(
      start = dn_info[["start"]], 
      end   = dn_info[["end"]], 
      col_names = dn_info[["names"]]),
  col_types = dn_info[["types"]])

}
