% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evnt_key.R
\name{get_evnt_key}
\alias{get_evnt_key}
\title{Extract event type from MEPS event dataset}
\usage{
get_evnt_key(df)
}
\arguments{
\item{df}{MEPS event dataset}
}
\value{
Two-character string of event key. If key not recognized, returns NULL.
}
\description{
Extracts event type based on two-character prefix for variable names on MEPS event datasets.
}
\examples{
dat <- read_MEPS(file='h102b',web=T)
get_evnt_key(dat)

}
