% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year.R
\name{get_year}
\alias{get_year}
\title{Extract data year from dataset}
\usage{
get_year(df)
}
\arguments{
\item{df}{MEPS dataset}
}
\value{
two-digit character key for year
}
\description{
Extracts data year based on the two-digit number in PERWT**F or WTDPER** (WTDPER** was converted to PERWT**F in 1999). The weight variable must be in the dataset.
}
\examples{

# Get data year from full year file
FYC <- read_MEPS(file='h147')
get_year(FYC)

# From event file
RX <- read_MEPS(file='h10a')
get_year(RX)
}
