% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MEPS.R
\name{read_MEPS}
\alias{read_MEPS}
\title{Import MEPS public use files into R as data frame or tibble}
\usage{
read_MEPS(file, ...)
}
\arguments{
\item{file}{name of public use file. Must be in standard format (e.g.
'h160g'). If 'file' is specified, it overrides any additional parameters.}

\item{...}{other parameters passed to the get_puf_names() function. Must
specify one of the following (see examples below): \cr \cr
\itemize{
  \item 'year' and 'type' (for most annual files)
  \item 'panel' with 'type' = "LONG" (for Longitudinal files)
  \item 'type' only if type = BRR or PL
 }}
}
\value{
MEPS data as a data frame or tibble.
}
\description{
This function reads in MEPS public use files (PUFs) from the MEPS website,
and imports them into R. Larger files (e.g.full-year-consolidated files) can
take several seconds to load. Internet connection required.
}
\examples{
# Use file name
OP2020 <- read_MEPS(file = "h220f")

# Use year and file type
OP2020 <- read_MEPS(year = 2020, type = "OP")

# Get Longitudinal file
LONG <- read_MEPS(type = "LONG", panel = 23)

# Get 4-year Longitudinal file (available for Panels 23 and 24)
LONG_4yr <- read_MEPS(type = "LONG", panel = 23, long_type = "4-year")

# Special cases: BRR file and Pooled Linkage
BRR <- read_MEPS(type = "BRR")
PL  <- read_MEPS(type = "PL")
}
