% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_evnt_key.R
\name{rm_evnt_key}
\alias{rm_evnt_key}
\title{Rename Variables of Event Data}
\usage{
rm_evnt_key(df, evnt_key)
}
\arguments{
\item{df}{data frame of MEPS event file (e.g. h160a for RX events from 2013)}

\item{evnt_key}{optional argument specifying event key to be removed from column names. Defaults to most common
2-letter prefix of current variable names}
}
\description{
This function renames columns of the event files to remove the event key prefix
(e.g. 'OBXP13X' => 'XP13X'), to facilitate stacking event files.
}
\examples{
# Read in office-based medical events file from 2013 ('h160g')
OB2013 = downloadMEPS('h160g')
OBnew = rm_evnt_key(OB2013)
head(OBnew)
}
