% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_events.R
\name{stack_events}
\alias{stack_events}
\title{Stack event files}
\usage{
stack_events(..., keep.vars = NULL, keep.all = F)
}
\arguments{
\item{...}{MEPS event datasets to stack.}

\item{keep.vars}{Vector of variables in addition to shared variables to keep in the stacked dataset.}

\item{keep.all}{If TRUE, keep all variables from all datasets. Can slow down computation.}
}
\value{
data frame of stacked event files with standardized variable names. Name of dataset is included as 'data' variable in stacked data set.
}
\description{
Stack MEPS event files. If needed, will remove event key prefixes to standardize variable names and will add total source of payment variables to hospital event files. By default, keeps only variables that are shared by the datasets (plus LINKIDX / EVNTIDX for ID variables)
}
\examples{

# Get event datasets
RX <- read_MEPS(year=2013,type='RX',web=T)
OB <- read_MEPS(year=2013,type='OB',web=T)
OP <- read_MEPS(year=2013,type='OP',web=T)
ER <- read_MEPS(year=2013,type='ER',web=T)

aa <- stack_events(RX,OB,OP,ER); head(aa);

# Force 'SEEDOC' and 'SEETLKPV' into stacked dataset
bb <- stack_events(RX,OB,OP,ER,keep.vars=c('SEEDOC','SEETLKPV')); head(bb);

}
