% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_meps.R
\name{dl_meps}
\alias{dl_meps}
\title{Download public use file from MEPS website}
\usage{
dl_meps(fname, ext = "dta", dir = tempdir())
}
\arguments{
\item{fname}{Name of public use file. Must be in standard format (e.g.
'h160g'). Can use the get_puf_names() function to look up file name by year
and type (requires internet connection).}

\item{ext}{Extension of file type. Options include "dta" (Stata data), "v9" (SAS V9 dataset), "xlsx" (Excel), "ssp" (SAS transport), and "dat" (ASCII). dta, v9, and xlsx are only available for data years 2018 and later.}

\item{dir}{(optional) Name of directory for file to be stored. If blank, will download to temporary folder}
}
\value{
Location and name of downloaded data file
}
\description{
This function downloads MEPS public use files (PUFs) from the MEPS website,
and saves them to a local directory (if specified)
}
\examples{
## Download MEPS 2019 Dental visits file from MEPS website

# Stata format (.dta) in "C:/MEPS"
meps_file <- dl_meps(fname = "h213b", ext = "dta", dir = "C:/MEPS")
haven::read_dta(meps_file) \%>\% head

# SAS V9 file in temporary directory
meps_file2 <- dl_meps(fname = "h213b", ext = "v9")
haven::read_sas(meps_file2) \%>\% head
}
