% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puf_names.R
\name{get_puf_names}
\alias{get_puf_names}
\title{Get MEPS Public Use File Names}
\usage{
get_puf_names(year, type, web = T)
}
\arguments{
\item{year}{(optional) Data year, between 1996 and most current PUF release. If omitted, files from all years will be returned}

\item{type}{(optional) Type of desired MEPS file. Options are: \cr \cr
Main files:
\itemize{
  \item "PIT" (Point-in-time file)
  \item "FYC" (Full-year consolidated file)
  \item "COND", "Conditions" (Medical conditions file)
  \item "Jobs" (Jobs file)
  \item "PRPL" (Person-round-plan file)
  \item "LONG", "Longitudinal"
 } \cr

 Event files:
 \itemize{
  \item "RX", "PMED" (Prescribed medicines)
  \item "Dental", "DV", "DN" (Dental visits)
  \item "Other_Medical", "OM" (Other medical expenses)
  \item "Inpatient", "IP","HS" (Hospital inpatient stays)
  \item "Emergency_Room", "ER" (Emergency room visits)
  \item "Outpatient", "OP" (Outpatient visits)
  \item "Office_based", "OB", "MV (Office-based medical provider visits)
  \item "Home_Health", "HH" (Home health)
  \item "CLNK" (Condition-event linkage file)
  \item "RXLK" (PMED-event linkage file)
} \cr

Other files:
\itemize{
  \item "Multum" (Multum Lexicon addendum files, 1996-2013)
  \item "PSAQ" (Preventative care SAQ, 2014)
  \item "MOS" (Medical Organizations Survey, 2015-2016)
  \item "FS" (Food Security file, 2016-2017)
  \item "BRR" (Balanced Repeated Replicates file)
  \item "PL", "Pooled Linkage" (Pooled Linkage file for common variance)
}}

\item{web}{if TRUE, returns names of .zip files from web, otherwise, returns names of .ssp files after download}
}
\description{
This is a lookup function that returns a single requested file name or list of names for specified MEPS data file.
}
\examples{
## Get file name for full-year-consolidated (FYC) file from 2005
get_puf_names(2005,'FYC')

## Get file names for all PUFs in 2014
get_puf_names(2014)

## Get file names for PMED event files, all years
get_puf_names(type='PMED')

## Return all files, all years
get_puf_names()

## Compare names of .ssp files with those on website links
get_puf_names(year = 1996, type = 'DV')
get_puf_names(year = 1996, type = 'DV', web = FALSE)
}
