% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puf_names.R
\name{get_puf_names}
\alias{get_puf_names}
\title{Get MEPS Public Use File Names}
\usage{
get_puf_names(year, type, download = T)
}
\arguments{
\item{year}{(optional) Data year, between 1996 and most current PUF release. If omitted, files from all years will be returned}

\item{type}{(optional) File type of desired MEPS file. Options are 'PIT' (Point-in-time file), 'FYC' (Full-year consolidated), 'Conditions' (Conditions file), 'Jobs' (Jobs file), 'PRP' (Person-Round-Plan), 'RX' (Prescription Medicines Events), 'DV' (Dental Visits), 'OM' (Other medical events), 'IP' (Inpatient Stays), 'ER' (Emergency Room Visits), 'OP' (Outpatient Visits), 'OB' (Office-based visits), 'HH' (Home health), 'CLNK' (conditions-event link file), and 'RXLK' (RX - events link file)}
}
\description{
This is a lookup function that returns a single requested file name or list of names for specified MEPS data file
}
\examples{
## Get file name for full-year-consolidated (FYC) file from 2005 
get_puf_names(2005,'FYC')

## Get file names for all PUFs in 2014
get_puf_names(2014)

## Get file names for RX event files, all years
get_puf_names(type='RX')

## Return all files, all years 
get_puf_names()
meps_names # can also just view meps_names

## Comapre names of files on website and .ssp files after downloading
get_puf_names(year = 1996, type = 'DV')
get_puf_names(year = 1996, type = 'DV', download=F)
}
